refactor(winner): rewrite Winner logic in a more idiomatic Kotlin style
- move tests each to test domain directory
