+++ src/main/kotlin/racingcar/InputValidator.kt
@@ -1,4 +1,4 @@
-package racingcar.io
+package racingcar
 
 object InputValidator {
