+++ src/main/kotlin/racingcar/domain/Winner.kt
@@ -2,18 +2,9 @@ package racingcar.domain
 
 object Winner {
     fun findWinner(cars: List<Car>): List<String> {
-        var max = 0
-        var winners = mutableListOf<Car>()
-        for (car in cars) {
-            if (max == car.getPosition()) {
-                winners.add(car)
-            }
-            if (max < car.getPosition()) {
-                winners = mutableListOf<Car>()
-                winners.add(car)
-                max = car.getPosition()
-            }
-        }
-        return winners.stream().map { car -> car.getName() }.toList()
+        val maxPosition = cars.maxOfOrNull { it.getPosition() } ?: return emptyList()
+        return cars
+            .filter { it.getPosition() == maxPosition }
+            .map { it.getName() }
     }
 }
