+++ src/main/kotlin/racingcar/io/InputHandler.kt
@@ -1,6 +1,7 @@
 package racingcar.io
 
 import camp.nextstep.edu.missionutils.Console
+import racingcar.InputValidator
 
 class InputHandler {
