+++ src/test/kotlin/racingcar/domain/RaceGameTest.kt
@@ -1,11 +1,8 @@
-package racingcar
+package racingcar.domain
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import racingcar.domain.Car
-import racingcar.domain.Engine
-import racingcar.domain.RaceGame
 import racingcar.io.Printer
 
 class RaceGameTest {
