+++ src/test/kotlin/racingcar/domain/WinnerTest.kt
@@ -1,10 +1,8 @@
-package racingcar
+package racingcar.domain
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
-import racingcar.domain.Car
-import racingcar.domain.Winner
 
 class WinnerTest {
